+++ apps/consent-ui/src/constants/index.ts
@@ -16,13 +16,4 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-import { AxiosError, AxiosResponse } from 'axios';
-
-export interface APIResponse {
-	error: AxiosError | undefined;
-	isLoading: boolean;
-	response: AxiosResponse | undefined;
-}
-
 export * from './externalPaths';
-export * from './api';
